WIP - Added functions to serialize/deserialize, both those in Function properties and in Program ObjectMap. Deserialize gets an error though

+++ m8rscript/Program.h
@@ -45,7 +45,7 @@ class Object;
 class Function;
 class SystemInterface;
 
-typedef Id<uint32_t> ObjectId;
+typedef Id<uint16_t> ObjectId;
 
 class Program : public Function {
 public:
